added NatSpec docs

+++ src/BaseChainSetup.sol
@@ -5,95 +5,165 @@ import {console2} from "forge-std/console2.sol";
 import {CommonBase} from "forge-std/Base.sol";
 import {ERC20} from "solmate/tokens/ERC20.sol";
 
+/// @title BaseChainSetup
+/// @notice BaseChainSetup is a base contract for setting up chain specific
+/// information. It is intended to be used as a base contract for
+/// tests and scripts that need to interact with multiple chains.
+/// It also provides an abstraction to reuse the same code for
+/// both foundry tests, forknets, testnets, and mainnet.
 contract BaseChainSetup is CommonBase {
+    /// @title runtime
+    /// @notice runtime is used to determine which environment we're on.
+    /// It is set by the inheriting contract.
+    /// @dev runtime is set by the inheriting contract. possible values are
+    /// ENV_FORGE_TEST, ENV_FORK, ENV_TESTNET, ENV_MAINNET
     string private runtime;
 
+    /// @title ENV_FORGE_TEST
+    /// @notice ENV_FORGE_TEST is the runtime value for foundry unit-tests
     string constant ENV_FORGE_TEST = "forge-test";
+    /// @title ENV_FORK
+    /// @notice ENV_FORK is the runtime value for local anvil networks
     string constant ENV_FORK = "fork";
+    /// @title ENV_TESTNET
+    /// @notice ENV_TESTNET is the runtime value for testnets
     string constant ENV_TESTNET = "testnet";
+    /// @title ENV_MAINNET
+    /// @notice ENV_MAINNET is the runtime value for mainnet
     string constant ENV_MAINNET = "mainnet";
 
+    /// @title broadcasting
+    /// @notice global variable to track if we're broadcasting
+    /// @dev it's used when switching chains to stop broadcasting
     bool broadcasting = false;
 
+    /// @title forkLookup
+    /// @notice forkLookup is a mapping of chain name to forkId
+    /// @dev forkId is what's given back from vm.createFork
     mapping(string => uint256) forkLookup;
+
+    /// @title gasEthLookup
+    /// @notice gasEthLookup is a mapping of chain name to whether or not
+    ///         the chain uses eth as its gas token. Like rollups, for example.
     mapping(string => bool) gasEthLookup;
+    /// @title wethLookup
+    /// @notice wethLookup is a mapping of chain name to the weth address on that chain
     mapping(string => address) wethLookup;
+    /// @title wrappedLookup
+    /// @notice wrappedLookup is a mapping of chain name to the wrapped token on that chain
+    /// on eth chains, this is the same as wethLookup
     mapping(string => address) wrappedLookup;
+    /// @title chainIdLookup
+    /// @notice chainIdLookup is a mapping of chain alias to the chainId on that chain
+    /// chain alias is defined in foundry.toml
     mapping(string => uint256) chainIdLookup;
 
-    function wethBalance(
-        string memory chain,
-        address user
-    ) public view returns (uint) {
+    /// @title wethBalance
+    /// @notice wethBalance returns the weth balance of a user on a given chain
+    /// @param chain is the chain alias
+    /// @param user is the user address
+    /// @return the weth balance of the user on the chain
+    function wethBalance(string memory chain, address user) public view returns (uint256) {
         return ERC20(getWeth(chain)).balanceOf(user);
     }
 
-    function wrappedBalance(
-        string memory chain,
-        address user
-    ) public view returns (uint) {
+    /// @title wrappedBalance
+    /// @notice wrappedBalance returns the wrapped balance of a user on a given chain on eth chains,
+    /// this is the same as wethBalance
+    /// @param chain is the chain alias
+    /// @param user is the user address
+    /// @return the wrapped balance of the user on the chain
+    function wrappedBalance(string memory chain, address user) public view returns (uint256) {
         return ERC20(getWrapped(chain)).balanceOf(user);
     }
 
-    function getWeth(
-        string memory chain
-    ) public view returns (address payable) {
+    /// @title getWeth
+    /// @notice getWeth returns the weth address on a given chain
+    /// @param chain is the chain alias
+    /// @return the weth address on the chain
+    function getWeth(string memory chain) public view returns (address payable) {
         return payable(_getTokenForChain(chain, "weth", wethLookup));
     }
 
-    function getWrapped(
-        string memory chain
-    ) public view returns (address payable) {
+    /// @title getWrapped
+    /// @notice getWrapped returns the wrapped address on a given chain, on eth chains,
+    /// this is the same as getWeth
+    /// @param chain is the chain alias
+    /// @return the wrapped address on the chain
+    function getWrapped(string memory chain) public view returns (address payable) {
         return payable(_getTokenForChain(chain, "wrapped", wrappedLookup));
     }
 
-    function _getTokenForChain(
-        string memory chain,
-        string memory tokenName,
-        mapping(string => address) storage lookup
-    ) private view returns (address) {
+    /// @title _getTokenForChain
+    /// @notice _getTokenForChain is a helper function to get the token address for a given chain
+    /// @param chain is the chain alias
+    /// @param tokenName is the name of the token
+    /// @param lookup is the mapping to look up the token address
+    function _getTokenForChain(string memory chain, string memory tokenName, mapping(string => address) storage lookup)
+        private
+        view
+        returns (address)
+    {
         address token = payable(lookup[chain]);
-        require(
-            token != address(0),
-            string.concat("no ", tokenName, " found for chain: ", chain)
-        );
+        require(token != address(0), string.concat("no ", tokenName, " found for chain: ", chain));
         return token;
     }
 
+    /// @title isMainnet
+    /// @notice isMainnet returns true if we're running the script on mainnet
     function isMainnet() public returns (bool) {
         return vm.envOr("MAINNET", false) && strCompare(runtime, ENV_MAINNET);
     }
 
+    /// @title isTestnet
+    /// @notice isTestnet returns true if we're running the script on testnet
     function isTestnet() public returns (bool) {
         return vm.envOr("TESTNET", false) && strCompare(runtime, ENV_TESTNET);
     }
 
-    function strCompare(
-        string memory s1,
-        string memory s2
-    ) public pure returns (bool) {
-        return
-            keccak256(abi.encodePacked(s1)) == keccak256(abi.encodePacked(s2));
+    /// @title strCompare
+    /// @notice strCompare compares two strings
+    /// @param s1 is the first string
+    /// @param s2 is the second string
+    /// @return true if the strings are equal
+    function strCompare(string memory s1, string memory s2) public pure returns (bool) {
+        return keccak256(abi.encodePacked(s1)) == keccak256(abi.encodePacked(s2));
     }
 
+    /// @title isForgeTest
+    /// @notice isForgeTest returns true if we're running the script in foundry unit-tests
     function isForgeTest() public view returns (bool) {
         return strCompare(runtime, ENV_FORGE_TEST);
     }
 
+    /// @title isForkRuntime
+    /// @notice isForkRuntime returns true if we're running the script on a forknet
     function isForkRuntime() public view returns (bool) {
         return strCompare(runtime, ENV_FORK);
     }
 
+    /// @title setRuntime
+    /// @notice setRuntime sets the runtime, the possible values are:
+    /// ENV_FORGE_TEST, ENV_FORK, ENV_TESTNET, ENV_MAINNET
     function setRuntime(string memory _runtime) internal {
         runtime = _runtime;
     }
 
-    function _forkAlias(
-        string memory _chain
-    ) internal view returns (string memory) {
+    /// @title _forkAlias
+    /// @notice _forkAlias returns the fork alias for a given chain, if we're running on a forknet
+    /// there is a prefix of "fork_" added to the chain alias in those cases, this is to be able to
+    /// separate between rpc's defined for chains, as well as forknets in foundry.toml
+    /// @param _chain is the chain alias
+    function _forkAlias(string memory _chain) internal view returns (string memory) {
         return isForkRuntime() ? string.concat("fork_", _chain) : _chain;
     }
 
+    /// @title startImpersonating
+    /// @notice startImpersonating starts impersonating a given address on a given chain, this abstracts away
+    /// the differences between foundry unit-tests, forknets, testnets, and mainnet so that the same
+    /// code can be used for all of them.
+    /// @param _as is the address to impersonate
+    /// @dev it uses vm.startPrank for foundry unit-tests, and vm.startBroadcast for everything else
     function startImpersonating(address _as) internal {
         console2.log("impersonating as", _as);
         if (isForgeTest()) {
@@ -104,22 +174,27 @@ contract BaseChainSetup is CommonBase {
         }
     }
 
-    function configureChain(
-        string memory chain,
-        bool isGasEth,
-        uint256 chainId,
-        address weth
-    ) public {
+    /// @title configureChain
+    /// @notice an overloading of configureChain that defaults to using weth as the wrapped token
+    /// @param chain is the chain alias
+    /// @param isGasEth is true if the chain uses eth as its gas token
+    /// @param chainId is the chainId for the chain
+    /// @param weth is the weth address on the chain
+    function configureChain(string memory chain, bool isGasEth, uint256 chainId, address weth) public {
         configureChain(chain, isGasEth, chainId, weth, weth);
     }
 
-    function configureChain(
-        string memory chain,
-        bool isGasEth,
-        uint256 chainId,
-        address weth,
-        address wrapped
-    ) public {
+    /// @title configureChain
+    /// @notice configureChain configures the chain specific information for a given chain,
+    /// it's used to populate the mappings for wethlookup, wrappedLookup, chainIdLookup, and gasEthLookup
+    /// @param chain is the chain alias
+    /// @param isGasEth is true if the chain uses eth as its gas token
+    /// @param chainId is the chainId for the chain
+    /// @param weth is the weth address on the chain
+    /// @param wrapped is the wrapped token address on the chain
+    function configureChain(string memory chain, bool isGasEth, uint256 chainId, address weth, address wrapped)
+        public
+    {
         try vm.createFork(_forkAlias(chain)) returns (uint256 forkId) {
             forkLookup[chain] = forkId;
         } catch {}
@@ -133,6 +208,11 @@ contract BaseChainSetup is CommonBase {
         chainIdLookup[chain] = chainId;
     }
 
+    /// @title stopImpersonating
+    /// @notice stopImpersonating stops impersonating a given address on a given chain, this abstracts away
+    /// the differences between foundry unit-tests, forknets, testnets, and mainnet so that the same
+    /// code can be used for all of them.
+    /// @dev it uses vm.stopPrank for foundry unit-tests, and vm.stopBroadcast for everything else
     function stopImpersonating() internal {
         if (isForgeTest()) {
             vm.stopPrank();
@@ -142,6 +222,12 @@ contract BaseChainSetup is CommonBase {
         }
     }
 
+    /// @title switchTo
+    /// @notice switchTo switches to a given chain, this abstracts away the differences between foundry unit-tests,
+    /// forknets, testnets, and mainnet so that the same code can be used for all of them.
+    /// @param chain is the chain alias
+    /// @dev it uses vm.selectFork for everything, but ensures to stop broadcasting before switching forks in case we're
+    /// not in a unit-test, and start broadcasting again after switching forks
     function switchTo(string memory chain) internal {
         if (bytes(chain).length == 0) {
             revert("no chain specified");
@@ -152,7 +238,7 @@ contract BaseChainSetup is CommonBase {
             broadcasting = false;
         }
 
-        uint forkId = forkLookup[chain];
+        uint256 forkId = forkLookup[chain];
 
         vm.selectFork(forkId);
 
@@ -166,17 +252,21 @@ contract BaseChainSetup is CommonBase {
         }
     }
 
-    function dealTo(
-        string memory chain,
-        address user,
-        uint256 amount
-    ) internal returns (bool success) {
+    /// @title dealTo
+    /// @notice dealTo deals a given amount to a given user on a given chain, this abstracts away the differences
+    /// between foundry unit-tests, forknets, testnets, and mainnet so that the same code can be used for all of them.
+    /// In case of a unit-test, it uses vm.deal, otherwise it sends eth to the user from the deployer account.
+    /// @param chain is the chain alias
+    /// @param user is the user address
+    /// @param amount is the amount to deal
+    /// @return true if the deal was successful
+    function dealTo(string memory chain, address user, uint256 amount) internal returns (bool success) {
         success = true;
         switchTo(chain);
         if (isForgeTest()) {
             vm.deal(user, amount);
         } else {
-            (success, ) = user.call{value: amount}("");
+            (success,) = user.call{value: amount}("");
         }
     }
 }
